requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cbpro==1.1.4
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
lxml==4.8.0
multidict==6.0.2
Pillow==9.2.0
pycparser==2.21
pymongo==3.5.1
PyNaCl==1.5.0
python-dotenv==0.21.0
pytube==12.1.0
requests==2.13.0
six==1.10.0
sortedcontainers==2.4.0
websocket-client==0.40.0
yarl==1.8.1
